Seperate reloads when reporting previews infobar dismissal
When the previews infobar is dismissed, it would be valuable to track
reloads differently from new navigations. Reloads are not treated as
opt outs by the blacklisting policy, so the page could be served as a new LoFi page or LitePage after a reload
(offline previews are specifically prevented from being shown on any
reload however).

If the user dismisses the infobar and then reloads, this will only be captured as the user dismissing the infobar.

BUG=672634

Review-Url: https://codereview.chromium.org/2557113004
Cr-Original-Commit-Position: refs/heads/master@{#439895}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0be759f29bd8acd4b83da1aa3e49c645da01ef3