Add page-level aggregation of SubresourceFilter time metrics.
Add a new message (SubresourceFilterHostMsg_DocumentLoadStatistics) that is sent
to a RenderFrameHost in the browser when a document load is finished, just
before the DidFinishLoad message, if performance measurements were enabled for
the load. Contains the total time spent on evaluating subresource loads in
DocumentSubresourceFilter::allowLoad() for a frame.

Measurements are aggregated in SubresourceFilterDriverFactory and recorded to
"SubresourceFilter.PageLoad.SubresourceEvaluation.*Duration" histograms when
DidFinishLoad is called for the main frame.

BUG=609747,672519

Review-Url: https://codereview.chromium.org/2581043003
Cr-Original-Commit-Position: refs/heads/master@{#440849}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 22be4914542ba9072ee2e6c7c3480881ebcaa6fe