Remove Render.Workers.MaxWorkerCountInRendererProcess UMA.
This UMA and code is misleading because the shared worker ref count in
RenderProcessHostImpl actually is not the number of running shared
workers. We've added comments since that clarify it.

If we want this UMA back we can add a new one later. I'll add current
stats to the bug.

BUG=403258

Review-Url: https://codereview.chromium.org/2607513002
Cr-Original-Commit-Position: refs/heads/master@{#440951}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3616a4304d02fdaba982fc443ed8b908cf786e78