UMA metric for LayerUpdateTimes
We are interested to learn about correlation between number of layers
and the cost of LayerUpdateTimes in actual sites to guide
some optimization efforts.

To do this, we collect stats on LayersUpdateTime but bucket that into
several suffixed histograms to be able to relate it to number of
layers.

BUG=661741

Review-Url: https://codereview.chromium.org/2484743002
Cr-Original-Commit-Position: refs/heads/master@{#441076}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 61be4cf0015eb2c0fb50cd58de61939423eab618