Add histograms for location timings for Android omnibox queries.
Add histograms for the duration between the obtaining of location and the use of
location for omnibox queries. One set of histograms records the time since the
location was first obtained. There is a histogram for each combination of
Location Source and whether a location was sent. A second set of histograms
records the time since the location was most recently obtained. There is a
histogram for each of the Location Sources.

BUG=668923

Review-Url: https://codereview.chromium.org/2576633002
Cr-Original-Commit-Position: refs/heads/master@{#441323}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b6572e3c0aba3d558e173440bd644b2e4b9bd04