Configure xctest build settings to prevent Xcode from linking objects.
Test files need to be known to Xcode for proper indexing and for discovery of
tests function for XCTest, but the compilation is done via ninja, which causes
Xcode fail find object files when trying to link.

This CL configures the build settings of XCTest targets to trick Xcode to
blindly return success when trying to link object files.

BUG=614818

Review-Url: https://codereview.chromium.org/2594033002
Cr-Original-Commit-Position: refs/heads/master@{#441500}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e5894a9a070055aec683a83d50523df7f585f38