Add "Useful Logs and Debugging Info" section to perf sheriff doc.
We have lots of places with useful debugging info but it can be
hard to find and to know what is useful to look at. I think it
would be useful to add a section to the perf sheriff doc
dedicated to describing the logging info that we have set up.

Review-Url: https://codereview.chromium.org/2611183005
Cr-Original-Commit-Position: refs/heads/master@{#442294}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: abbb2541620aa14479081acce93c94c52523b6d5