Show the search geolocation disclosure from geolocation API use.
The new search geolocation system will automatically grant geolocation
API access to the default search engine. This change shows the
disclosure UI when this happens so the user is aware of it.

This change also fixes a bug with showing the API introduced by an
earlier change that prevented the disclosure from being shown on
omnibox queries.

BUG=674398

Review-Url: https://codereview.chromium.org/2627853002
Cr-Original-Commit-Position: refs/heads/master@{#443553}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3c6a28850a816162c59526005665c683ca42308