[NTP::SectionOrder] Add category position metric for opened suggestions.
For each opened suggestion record its category position (i.e. rank). The
same is done per each category as well. E.g. if a suggestion in category
A (which is located at position X) is clicked, X is emited into general
histogram and into A's historgam.

This is done in order to reflect dynamic category order and measure
how well it is performing.

BUG=676264

Review-Url: https://codereview.chromium.org/2609413005
Cr-Original-Commit-Position: refs/heads/master@{#444340}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6edff76757001e14e1ea38c1e4b4e9f1cdf74d32