[NTP Client] Use first seen position in suggestion UMA
Updates the description and calculation to be more consistent
with what we want to track and more accurate across the different
modifications the suggestion surface can go through.
This means that instead of tracking the position at the moment the
item is loaded, we track the position when it is bound to the view
and the user has a chance to see it.

BUG=675623

Review-Url: https://codereview.chromium.org/2640443003
Cr-Original-Commit-Position: refs/heads/master@{#444693}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ecb7bd1cd2294ead17cb71265c98a9574dd5f342