[Video] Fix the TimeFromForegroundToFirstFrame metric.
Only start tracking the foreground time if the video we had or would optimize is playing or will be playing when shown.

Thus we don't track the time for paused videos or videos we wouldn't optimize like streams so they don't improve the average results.

BUG=670150
TEST=manual

Review-Url: https://codereview.chromium.org/2631633003
Cr-Original-Commit-Position: refs/heads/master@{#444816}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc273ddad35337b82ddfdb77248986fc928d035e