predictors: Collect page load metrics for ResourcePrefetchPredictor.
This CL adds ResourcePrefetchPredictorPageLoadMetricsObserver that allows to
collect page load metrics only for prefetchable pages.
For now, the page is considered prefetchable if it has resources in the
predictor DB that could be used for prefetching. This is subject to change
depending on triggering selection for the trial.

BUG=680049

Review-Url: https://codereview.chromium.org/2632503005
Cr-Original-Commit-Position: refs/heads/master@{#445723}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 666d5ac67f9172f965581c314de225737ba95d7d