Remove Physical Web notifications
This change is a first pass in removing Physical Web notifications.
It's already the case that Physical Web notifications do not fire, but
this change removes the code path that would fire notifications.  There
is still more vestigial code that will be removed in later changes.

BUG=680747

Review-Url: https://codereview.chromium.org/2650793005
Cr-Original-Commit-Position: refs/heads/master@{#445784}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53466abaf91e1fe99c116cb275b4643be65ebcf5