[animations] Adds metrics for jank on selected layer animations
This CL is WIP to add UMA metrics for "smoothness" factor on
selected layer animations. The UI developer would be able to
tag selected animations (sequences or elements) with a "name"
and then have UMA data available for how smooth (how close to
ideal 60 fps) the animation is in the field.

BUG=684115

Review-Url: https://codereview.chromium.org/2631333002
Cr-Original-Commit-Position: refs/heads/master@{#446075}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62ba671daf03f09c08257fab4559a31259aa09ce