predictors: Add prefetcher histograms for speculative prefetch.
Prefetching is most useful when a prefetched resource isn't in cache or needs
to be revalidated. This CL adds a histogram that allows us to estimate how
often the predictor prefetches resources from the network and how often from
the cache. This CL also adds histograms for a single prefetcher instance:
count of prefeched resources and amount of consumed KB.

BUG=680049

Review-Url: https://codereview.chromium.org/2654913004
Cr-Original-Commit-Position: refs/heads/master@{#446309}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 075c5d682d0b00fb8e311a23990082567c8b98e3