Substitute boolean GDI printing for a mode type
Instead of indicating GDI printing with a boolean add a parameter to the
PdfRenderSettings structure that is an enum. This makes a few function
calls cleaner and will facilitate addition of new modes when postscript
printing is added (see https://codereview.chromium.org/2633573002/).

BUG=None

Review-Url: https://codereview.chromium.org/2640033005
Cr-Original-Commit-Position: refs/heads/master@{#446484}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9251407fa9abe958dcbf2d72e7c25f8a422542c