Update static_initializers tool to build easier, find more
static_initializers failed to find the initializer/destructor code for
g_fallback_crash_handler_launcher because it was of this form:
    `dynamic atexit destructor for '

So, this pattern is now watched for. This change also adds a build.bat
file and .gitignore file. If we add this to the bots we'll need to make
it part of the .gn build.

R=scottmg@chromium.org
BUG=677351

Review-Url: https://codereview.chromium.org/2656223002
Cr-Original-Commit-Position: refs/heads/master@{#446524}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4419391c9bd15c97e01fcacbf40871344476b5ca