Add Desktop iOS promotion logging to chrome ios app.
Create a sync observer for the desktop promotion and add it
on the application main controller.
When the priority prefs are synced. log to UMA if the user have seen the
desktop promotion on the last 7 days and log if they triggered the sms feature
on the promotion.
Also mark the promotion as completed using priority prefs.

BUG=681885

Review-Url: https://codereview.chromium.org/2643723004
Cr-Original-Commit-Position: refs/heads/master@{#447026}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54a832541cb8aafced2f60deb296132c90ad1eb4