Add a policy to always show the Cast toolbar icon
This CL adds a policy "ShowCastIconInToolbar," which will force the media router
toolbar action to be always shown in the toolbar or the overflow menu when
enabled.

When the policy is enabled, the action's context menu item "Always show icon"
will be replaced by "Added by your administrator," similar to the item
"Installed by your administrator" for policy-enabled extensions. This menu item
will always be disabled.

This CL adds a new pref with ID prefs::kShowCastIconInToolbar for the policy.
When we later remove ComponentMigrationHelper and the prefs associated with it
that we currently use to pin the media router action to the toolbar, we can use
prefs::kShowCastIconInToolbar to record user-initiated pinning of the action to
the toolbar.

BUG=666003

Review-Url: https://codereview.chromium.org/2559323002
Cr-Original-Commit-Position: refs/heads/master@{#447289}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a5f2851a8a4cf5222a676dc8bcc7d52fc29a7f1