Record UMA histogram when Form-Not-Secure warning is shown
This records a histogram, at most once per main-frame navigation, when "Login
not secure" is shown in a password form. This will allow us to measure the
frequency of the warning (e.g. the number of times the warning is shown per
million navigations.)

BUG=687823

Review-Url: https://codereview.chromium.org/2670733002
Cr-Original-Commit-Position: refs/heads/master@{#447960}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce9ab459b6cce4f6d327b83ef49a8684a8839b9b