Remove Media user actions.
User actions should be used only if timestamping is important. We suggest using enum histogram if we just want the number of times this has occured, which takes advantage of aggregation.
If these are important, let me know and I can switch it to a histogram.

BUG=

Review-Url: https://codereview.chromium.org/2679543006
Cr-Original-Commit-Position: refs/heads/master@{#448766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4611e92629d2721cf402c10cf0c3c62d8e21ff3