Re-enable media.android.tough_video_cases.
This was disabled in cr/2402083002 and  cr/2549373003.

The problem was that we became idle during seek, so we would suspend if the seek
took too long. sandersd@ fixed this in cr/2490783002.

liberato@ tested this both right before and right after sandersd@'s
change. Before the change it would usually fail but sometimes succeed.
After the change it would always succeed.

BUG=647372

Review-Url: https://codereview.chromium.org/2684033009
Cr-Original-Commit-Position: refs/heads/master@{#449538}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4f0e124c5a8884a16459c67d956d011be8acc5a