Integration test of Data Saver block-once directive
The added test checks that a test HTML resource is not served
via the Data Saver proxy and that a subresource referenced in
the HTML is served via the proxy. The proxy is configured to
respond with the 'block-once' directive (a field of the
Chrome-Proxy header) if the test HTML resource is requested.

BUG=680574

Review-Url: https://codereview.chromium.org/2693873004
Cr-Original-Commit-Position: refs/heads/master@{#450142}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f40be110899bd6c3a4dc9a90daa10ba8adfe12c