Export the first load latency of the virtual keyboard.
The metric FirstLoadTime was not listed in histograms.xml.
I renamed FirstLoadTime to InitLatency.FirstLoad to be consist with
other latency metrics named InputMethod.VirtualKeyboard.InitLatency.*,
and exported it. To distinguish the metrics on JS and on C++, the prefix
"InputMethod." was intentionally omitted.

BUG=None
Test=manually confirmed the metric VirtualKeyboard.InitLatency.FirstLoad
was exported in chrome://histograms using Chrome on Linux.

Review-Url: https://codereview.chromium.org/2684993010
Cr-Original-Commit-Position: refs/heads/master@{#450161}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 797648acb30247ac0d91087e86828fb3b905dca9