Remove the notificationProvider extension API
The NotificationProvider API was meant to provide a way for Chrome Apps
and extensions to take over handling of notifications from the Message
Center, for example to defer to the native notification center on a
system. However, the API was neither completed, nor shipped.

Today we're working on implementing support for native notification
centers directly to Chrome, not requiring any additional extension
support. This is already supported on Mac and is in development for
Windows. This makes an important use-case of the API redundant.

Let's remove the API. We can always revisit in the future.

BUG=397197, 482733

Review-Url: https://codereview.chromium.org/2659533003
Cr-Original-Commit-Position: refs/heads/master@{#450333}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f9fbe22a90c39cb811e94976923ebff0bc7f939f