Update ThreatDetails code to populate the DOM in CSBRR report.
The complexity here surrounds mapping the nodes from any particular render frame up to the specific iframe element in the parent frame that those nodes belong to. Currently this is done by matching the src attribute of the iframe element to the URL of the render frame. In some cases, the URL of the render frame can be absent (eg: about:blank). Not sure how often this will happen in the wild, though, so mostly just tracking how often this case occurs for now.

BUG=674618

Review-Url: https://codereview.chromium.org/2693483002
Cr-Original-Commit-Position: refs/heads/master@{#451058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 685b83906193d490f79a315b2c2a021d919d1e84