Add a use counter for RtcpMuxPolicy of "negotiate".
When parsing the RTCConfiguration, if the RtcpMuxPolicy is "negotiate",
it will be counted. The number will be used to estimate the risk of
removing this feature.

Intent to deprecate thread: https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/OP2SGSWF5lo

BUG=chromium:685727

Review-Url: https://codereview.chromium.org/2679093006
Cr-Original-Commit-Position: refs/heads/master@{#451162}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 887a0d5ac980c21eebecb133a75e08dc2fe048bf