Add option to GRIT to provide a resource ordering input file.
This allows optimizing Chrome's start up sequence such that the
resources that are loading during startup appear in order in the
resource .pak files and thus incur less page faults to load from
disk and a faster start up time and less memory used (due to less
pages resident in RAM).

This change adds the necessary infrastructure to support this,
but does not yet enable it for any platform. That will be done in
a separate change later.

BUG=692670

Review-Url: https://codereview.chromium.org/2690263004
Cr-Original-Commit-Position: refs/heads/master@{#451464}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a84244d0aabd3d3ee775adf82dc9acb2fefac31