Implement permission embargo suppression metrics.
This CL adds a new UMA histogram to record whether or not permission
prompts are suppressed due to embargo. This allows us to measure how
effective embargo is at reducing the number of prompts seen by users.

New histogram tests are implemented to verify that the metrics are
recorded as expected.

BUG=679877

Review-Url: https://codereview.chromium.org/2701343002
Cr-Original-Commit-Position: refs/heads/master@{#452466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2391315fa51475da9ea5314ed81990e709644659