Add Windows-specific WPR archives for YouTube system health stories
Even with a platform-specific recording, this only kind of works. It
seems like all videos still stall around 10 seconds into the video.
Given that the YouTube team surprised that capturing the net stream
of the site and playing it back worked at all, I think this is as good
as can be expected for now.

BUG=618698

Review-Url: https://codereview.chromium.org/2718473002
Cr-Original-Commit-Position: refs/heads/master@{#452657}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 106c52c87b9deeb8be9127466bec935b695369d3