color: Update ICC histograms based on results
Based on the data gathered with the previous set of histograms, we can
assume that the range of the transfer function is [0, 1]. Bake this
assumption into the numerical approximation model (so we're solving for
fewer parameters).

Merge some of the histograms together under ICC analysis.

Add a histogram of the combined channels' convergence results and error
as well.

BUG=667966

Review-Url: https://codereview.chromium.org/2717263003
Cr-Original-Commit-Position: refs/heads/master@{#453645}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3aa7226ee993cda646b50b271753d3bdf661cf9