Simplify page foreground duration tracking.
In https://codereview.chromium.org/2699933003 we added experimental metrics to
track cases of time durations where a page spends time in the foreground: time
to first background, time to page 'end', and time until Chrome is backgrounded,
on Android, where we may be killed once Chrome goes into the background.

UMA shows that the histograms for these 3 cases are comparable in shape and
don't have any strange outliers. So we can simplify by merging the 3 histograms
into a single overall histogram that measures the period that a page spends in
the foreground.

This change merges these 3 metrics into a single high level metric that tracks
the time a page spends in the foreground. This metric is intended to help us
understand how much time users are spend engaging with web pages in Chrome,
which can be used as a companion to the existing 'time in browser' metric as
well as to understand if field trials lead to increased time spent on web pages.

This change also promotes navigation_start to a field in PageLoadExtraInfo - we
need this to provide a common implementation for computing foreground duration,
and multiple observers want this field so it makes sense to provide it in a
single common place rather than requiring observers to retrieve it from the nav
handle.

BUG=693725

Review-Url: https://codereview.chromium.org/2719183005
Cr-Original-Commit-Position: refs/heads/master@{#453815}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9f47e5e98c9f7745c8d70bfa36cfa647397dec0