[Remote suggestions] Move the throttler into the Scheduler
This CL moves the throttler into the Scheduler -- which is now the
central place to decide whether to and when to fetch.

This change also helps in employing I/O scheduler in the Scheduler for
prioritization of background fethes. I/O scheduler tasks cannot run on
the main thread and thus cannot use PrefService. As throttler depends
on prefs, the easiest way is to ask for quota before posting the task
to I/O scheduler.

This change removes the quota_error fetch status reporting from UMA.
This is no problem as all the quota requests are reported to UMA from
the throttler anyway.

BUG=694325

Review-Url: https://codereview.chromium.org/2714863004
Cr-Original-Commit-Position: refs/heads/master@{#454258}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a06fdaf8ba732ae795bcb1f2331523babc88dd8