Deprecated the --allow-no-sandbox-job flag in favor of automatic recognition.
Instead of requiring the user to set the flag when running Chrome in TS
environment, try to recognize the environment and allow the sandbox to
run without this flag whenever needed.

Measure how often the automatic recognition for terminal services environment would have incorrectly decided that the job object should be applied when it shouldn't have been as dictated by the flag --allow-no-sanbox-job.

BUG=687147
TEST=Manual in TS env on Win Srv 2k8 R2.

Review-Url: https://codereview.chromium.org/2672003002
Cr-Original-Commit-Position: refs/heads/master@{#454429}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7091f0852e6a60119f3cdbcc7eb34c479bf93fc9