[Android Crash Reporting] Simplify crash report upload code.
In particular: run the code to extract logcat output within the context
of the main Chrome process/service, rather than kicking off the work to
a pair of background services.  Moreover, dramatically simplify the
MetricsUploadService code: remove all but one of the intents that it
handles.

BUG=694884
R=mariakhomenko@chromium.org

Review-Url: https://codereview.chromium.org/2727573004
Cr-Original-Commit-Position: refs/heads/master@{#455704}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a9a8f4e062bd1c72c3b511de3c41611e24b7949e