This tries to use the old buildbot endpoints, which have since been deprecated. Switching this to use the new milo endpoint, specifically the GetCompressedMasterJSON (https://github.com/luci/luci-go/blob/master/milo/api/proto/buildbot.proto).
A lot of this was lifted straight from:
https://cs.chromium.org/chromium/infra/appengine/findit/waterfall/buildbot.py

There's also the CBE master endpoints which give the same data roughly. The code is simpler but I found it was taking a lot longer to come back so I opted for the milo endpoint.

Review-Url: https://codereview.chromium.org/2752573003
Cr-Original-Commit-Position: refs/heads/master@{#456942}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7005dc9fffe6f00c583033892ef07779793834d2