[Sync] Add USER_EVENTS model type.
This does not register the type or instantiate any new objects. The
only change is that the type exists, and will be tied to the history
checkbox in sync settings.

Also organized a few model type listing that were near alphabetical,
but not quite.

BUG=701032

Review-Url: https://codereview.chromium.org/2692923012
Cr-Original-Commit-Position: refs/heads/master@{#457565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65f1ba166c45456df7947ef9a8888cefcee313fa