Add UMA for how long service workers run for.
We don't have any UMA for how long a SW stays alive, so this
is a step toward fixing that. This adds UMA:
* ServiceWorker.Runtime
  - How long the worker ran for (wall time). Recorded when
a service worker stops.

Ideally we'd additionally like to know 1) about SW's that
never stop, and 2) why a SW stays alive particularly
if it's for a long time, and also be able to distinguish
valid use from abusive use, but there's some complexity
there so we're starting simple.

BUG=683047

Review-Url: https://codereview.chromium.org/2706923003
Cr-Original-Commit-Position: refs/heads/master@{#457669}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f93cb21f636207e7f0573bf99681bfce4951e0f7