Implement the new Photo picker, part one.
This is the first step towards implementing a new Photo Picker for Chrome for
Android, broken down into multiple steps for easier reviewing.

First part includes simply replacing the stock Android picker with a Toast
saying "Not implemented".

Includes a Finch tie-in and a flag to turn this on manually ("Enable new
photopicker" in chrome://flags).

BUG=656015

Review-Url: https://codereview.chromium.org/2737393002
Cr-Original-Commit-Position: refs/heads/master@{#457741}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a6a921c6657a051d484306088a1d85b992f1a64f